@@ -2,11 +2,22 @@ import {path} from 'ramda';
 import {isReady} from '@plotly/dash-component-plugins';
 
 import Registry from '../registry';
+import {getPath} from './paths';
+import {stringifyId} from './dependencies';
 
 export default (layout, paths, targets) => {
+    if (!targets.length) {
+        return true;
+    }
     const promises = [];
+
+    const {events} = paths;
+    const rendered = new Promise(resolveRendered => {
+        events.once('rendered', resolveRendered);
+    });
+
     targets.forEach(id => {
-        const pathOfId = paths[id];
+        const pathOfId = getPath(paths, id);
         if (!pathOfId) {
             return;
         }
@@ -20,7 +31,14 @@ export default (layout, paths, targets) => {
         const ready = isReady(component);
 
         if (ready && typeof ready.then === 'function') {
-            promises.push(ready);
+            promises.push(
+                Promise.race([
+                    ready,
+                    rendered.then(
+                        () => document.getElementById(stringifyId(id)) && ready
+                    ),
+                ])
+            );
         }
     });
 
