@@ -4,6 +4,4 @@ from . import development  # noqa: F401
 from . import exceptions  # noqa: F401
 from . import resources  # noqa: F401
 from .version import __version__  # noqa: F401
-from ._callback_context import CallbackContext as _CallbackContext
-
-callback_context = _CallbackContext()
+from ._callback_context import callback_context  # noqa: F401
