@@ -7,12 +7,17 @@ import hashlib
 import collections
 import subprocess
 import logging
-from io import open  # pylint: disable=redefined-builtin
+import io
+import json
 from functools import wraps
 import future.utils as utils
 from . import exceptions
 
 logger = logging.getLogger()
+
+# py2/3 json.dumps-compatible strings - these are equivalent in py3, not in py2
+# note because we import unicode_literals u"" and "" are both unicode
+_strings = (type(""), type(utils.bytes_to_native_str(b"")))
 
 
 def interpolate_str(template, **data):
@@ -33,9 +38,7 @@ def format_tag(tag_name, attributes, inner="", closed=False, opened=False):
         tag += ">" + inner + "</{tag}>"
     return tag.format(
         tag=tag_name,
-        attributes=" ".join(
-            ['{}="{}"'.format(k, v) for k, v in attributes.items()]
-        ),
+        attributes=" ".join(['{}="{}"'.format(k, v) for k, v in attributes.items()]),
     )
 
 
@@ -56,46 +59,40 @@ def get_asset_path(requests_pathname, asset_path, asset_url_path):
 
 
 def get_relative_path(requests_pathname, path):
-    if requests_pathname == '/' and path == '':
-        return '/'
-    elif requests_pathname != '/' and path == '':
+    if requests_pathname == "/" and path == "":
+        return "/"
+    elif requests_pathname != "/" and path == "":
         return requests_pathname
-    elif not path.startswith('/'):
+    elif not path.startswith("/"):
         raise exceptions.UnsupportedRelativePath(
-            "Paths that aren't prefixed with a leading / are not supported.\n" +
-            "You supplied: {}".format(path)
+            "Paths that aren't prefixed with a leading / are not supported.\n"
+            + "You supplied: {}".format(path)
         )
-    return "/".join(
-        [
-            requests_pathname.rstrip("/"),
-            path.lstrip("/")
-        ]
-    )
+    return "/".join([requests_pathname.rstrip("/"), path.lstrip("/")])
+
 
 def strip_relative_path(requests_pathname, path):
     if path is None:
         return None
-    elif ((requests_pathname != '/' and
-            not path.startswith(requests_pathname.rstrip('/')))
-            or (requests_pathname == '/' and not path.startswith('/'))):
+    elif (
+        requests_pathname != "/" and not path.startswith(requests_pathname.rstrip("/"))
+    ) or (requests_pathname == "/" and not path.startswith("/")):
         raise exceptions.UnsupportedRelativePath(
-            "Paths that aren't prefixed with a leading " +
-            "requests_pathname_prefix are not supported.\n" +
-            "You supplied: {} and requests_pathname_prefix was {}".format(
-                path,
-                requests_pathname
+            "Paths that aren't prefixed with a leading "
+            + "requests_pathname_prefix are not supported.\n"
+            + "You supplied: {} and requests_pathname_prefix was {}".format(
+                path, requests_pathname
             )
         )
-    if (requests_pathname != '/' and
-            path.startswith(requests_pathname.rstrip('/'))):
+    if requests_pathname != "/" and path.startswith(requests_pathname.rstrip("/")):
         path = path.replace(
             # handle the case where the path might be `/my-dash-app`
             # but the requests_pathname_prefix is `/my-dash-app/`
-            requests_pathname.rstrip('/'),
-            '',
-            1
+            requests_pathname.rstrip("/"),
+            "",
+            1,
         )
-    return path.strip('/')
+    return path.strip("/")
 
 
 # pylint: disable=no-member
@@ -163,12 +160,54 @@ def create_callback_id(output):
     if isinstance(output, (list, tuple)):
         return "..{}..".format(
             "...".join(
-                "{}.{}".format(x.component_id, x.component_property)
+                "{}.{}".format(
+                    # A single dot within a dict id key or value is OK
+                    # but in case of multiple dots together escape each dot
+                    # with `\` so we don't mistake it for multi-outputs
+                    x.component_id_str().replace(".", "\\."),
+                    x.component_property,
+                )
                 for x in output
             )
         )
 
-    return "{}.{}".format(output.component_id, output.component_property)
+    return "{}.{}".format(
+        output.component_id_str().replace(".", "\\."), output.component_property
+    )
+
+
+# inverse of create_callback_id - should only be relevant if an old renderer is
+# hooked up to a new back end, which will only happen in special cases like
+# embedded
+def split_callback_id(callback_id):
+    if callback_id.startswith(".."):
+        return [split_callback_id(oi) for oi in callback_id[2:-2].split("...")]
+
+    id_, prop = callback_id.rsplit(".", 1)
+    return {"id": id_, "property": prop}
+
+
+def stringify_id(id_):
+    if isinstance(id_, dict):
+        return json.dumps(id_, sort_keys=True, separators=(",", ":"))
+    return id_
+
+
+def inputs_to_dict(inputs_list):
+    inputs = {}
+    for i in inputs_list:
+        inputsi = i if isinstance(i, list) else [i]
+        for ii in inputsi:
+            id_str = stringify_id(ii["id"])
+            inputs["{}.{}".format(id_str, ii["property"])] = ii.get("value")
+    return inputs
+
+
+def inputs_to_vals(inputs):
+    return [
+        [ii.get("value") for ii in i] if isinstance(i, list) else i.get("value")
+        for i in inputs
+    ]
 
 
 def run_command_with_process(cmd):
@@ -186,7 +225,7 @@ def run_command_with_process(cmd):
 
 
 def compute_md5(path):
-    with open(path, encoding="utf-8") as fp:
+    with io.open(path, encoding="utf-8") as fp:
         return hashlib.md5(fp.read().encode("utf-8")).hexdigest()
 
 
