@@ -11,10 +11,8 @@ import pkgutil
 import threading
 import re
 import logging
-import pprint
 
 from functools import wraps
-from textwrap import dedent
 
 import flask
 from flask_compress import Compress
@@ -23,23 +21,29 @@ from werkzeug.debug.tbtools import get_current_traceback
 import plotly
 import dash_renderer
 
-from .dependencies import Input, Output, State
 from .fingerprint import build_fingerprint, check_fingerprint
 from .resources import Scripts, Css
-from .development.base_component import Component, ComponentRegistry
-from . import exceptions
-from ._utils import AttributeDict as _AttributeDict
-from ._utils import interpolate_str as _interpolate
-from ._utils import format_tag as _format_tag
-from ._utils import generate_hash as _generate_hash
-from ._utils import patch_collections_abc as _patch_collections_abc
-from . import _watch
-from ._utils import get_asset_path as _get_asset_path
-from ._utils import create_callback_id as _create_callback_id
-from ._utils import get_relative_path as _get_relative_path
-from ._utils import strip_relative_path as _strip_relative_path
-from ._configs import get_combined_config, pathname_configs
+from .development.base_component import ComponentRegistry
+from .exceptions import PreventUpdate, InvalidResourceError
 from .version import __version__
+from ._configs import get_combined_config, pathname_configs
+from ._utils import (
+    AttributeDict,
+    create_callback_id,
+    format_tag,
+    generate_hash,
+    get_asset_path,
+    get_relative_path,
+    inputs_to_dict,
+    inputs_to_vals,
+    interpolate_str,
+    patch_collections_abc,
+    split_callback_id,
+    stringify_id,
+    strip_relative_path,
+)
+from . import _validate
+from . import _watch
 
 _default_index = """<!DOCTYPE html>
 <html>
@@ -67,15 +71,14 @@ _app_entry = """
 </div>
 """
 
-_re_index_entry = re.compile(r"{%app_entry%}")
-_re_index_config = re.compile(r"{%config%}")
-_re_index_scripts = re.compile(r"{%scripts%}")
-_re_renderer_scripts = re.compile(r"{%renderer%}")
+_re_index_entry = "{%app_entry%}", "{%app_entry%}"
+_re_index_config = "{%config%}", "{%config%}"
+_re_index_scripts = "{%scripts%}", "{%scripts%}"
 
-_re_index_entry_id = re.compile(r'id="react-entry-point"')
-_re_index_config_id = re.compile(r'id="_dash-config"')
-_re_index_scripts_id = re.compile(r'src=".*dash[-_]renderer.*"')
-_re_renderer_scripts_id = re.compile(r'id="_dash-renderer')
+_re_index_entry_id = 'id="react-entry-point"', "#react-entry-point"
+_re_index_config_id = 'id="_dash-config"', "#_dash-config"
+_re_index_scripts_id = 'src="[^"]*dash[-_]renderer[^"]*"', "dash-renderer"
+_re_renderer_scripts_id = 'id="_dash-renderer', "new DashRenderer"
 
 
 class _NoUpdate(object):
@@ -85,6 +88,13 @@ class _NoUpdate(object):
 
 # Singleton signal to not update an output, alternative to PreventUpdate
 no_update = _NoUpdate()
+
+
+_inline_clientside_template = """
+var clientside = window.dash_clientside = window.dash_clientside || {{}};
+var ns = clientside["{namespace}"] = clientside["{namespace}"] || {{}};
+ns["{function_name}"] = {clientside_function};
+"""
 
 
 # pylint: disable=too-many-instance-attributes
@@ -176,7 +186,7 @@ class Dash(object):
     :param index_string: Override the standard Dash index page.
         Must contain the correct insertion markers to interpolate various
         content into it depending on the app config and components used.
-        See https://dash.plot.ly/external-resources for details.
+        See https://dash.plotly.com/external-resources for details.
     :type index_string: string
 
     :param external_scripts: Additional JS files to load with the page.
@@ -221,7 +231,7 @@ class Dash(object):
         requests_pathname_prefix=None,
         routes_pathname_prefix=None,
         serve_locally=True,
-        compress=True,
+        compress=None,
         meta_tags=None,
         index_string=_default_index,
         external_scripts=None,
@@ -231,16 +241,7 @@ class Dash(object):
         plugins=None,
         **obsolete
     ):
-        for key in obsolete:
-            if key in ["components_cache_max_age", "static_folder"]:
-                raise exceptions.ObsoleteKwargException(
-                    key + " is no longer a valid keyword argument in Dash "
-                    "since v1.0. See https://dash.plot.ly for details."
-                )
-            # any other kwarg mimic the built-in exception
-            raise TypeError(
-                "Dash() got an unexpected keyword argument '" + key + "'"
-            )
+        _validate.check_obsolete(obsolete)
 
         # We have 3 cases: server is either True (we create the server), False
         # (defer server creation) or a Flask app instance (we use their server)
@@ -258,7 +259,7 @@ class Dash(object):
             url_base_pathname, routes_pathname_prefix, requests_pathname_prefix
         )
 
-        self.config = _AttributeDict(
+        self.config = AttributeDict(
             name=name,
             assets_folder=os.path.join(
                 flask.helpers.get_root_path(name), assets_folder
@@ -276,14 +277,12 @@ class Dash(object):
             routes_pathname_prefix=routes_prefix,
             requests_pathname_prefix=requests_prefix,
             serve_locally=serve_locally,
-            compress=compress,
+            compress=get_combined_config("compress", compress, True),
             meta_tags=meta_tags or [],
             external_scripts=external_scripts or [],
             external_stylesheets=external_stylesheets or [],
             suppress_callback_exceptions=get_combined_config(
-                "suppress_callback_exceptions",
-                suppress_callback_exceptions,
-                False,
+                "suppress_callback_exceptions", suppress_callback_exceptions, False
             ),
             show_undo_redo=show_undo_redo,
         )
@@ -306,8 +305,10 @@ class Dash(object):
             "via the Dash constructor"
         )
 
-        # list of dependencies
+        # list of dependencies - this one is used by the back end for dispatching
         self.callback_map = {}
+        # same deps as a list to catch duplicate outputs, and to send to the front end
+        self._callback_list = []
 
         # list of inline scripts
         self._inline_scripts = []
@@ -333,7 +334,7 @@ class Dash(object):
         self._cached_layout = None
 
         self._setup_dev_tools()
-        self._hot_reload = _AttributeDict(
+        self._hot_reload = AttributeDict(
             hash=None,
             hard=False,
             lock=threading.RLock(),
@@ -346,7 +347,7 @@ class Dash(object):
         self.logger = logging.getLogger(name)
         self.logger.addHandler(logging.StreamHandler(stream=sys.stdout))
 
-        if isinstance(plugins, _patch_collections_abc("Iterable")):
+        if isinstance(plugins, patch_collections_abc("Iterable")):
             for plugin in plugins:
                 plugin.plug(self)
 
@@ -380,65 +381,47 @@ class Dash(object):
             # gzip
             Compress(self.server)
 
-        @self.server.errorhandler(exceptions.PreventUpdate)
+        @self.server.errorhandler(PreventUpdate)
         def _handle_error(_):
             """Handle a halted callback and return an empty 204 response."""
             return "", 204
 
-        prefix = config.routes_pathname_prefix
-
         self.server.before_first_request(self._setup_server)
 
         # add a handler for components suites errors to return 404
-        self.server.errorhandler(exceptions.InvalidResourceError)(
-            self._invalid_resources_handler
-        )
-
-        self._add_url("{}_dash-layout".format(prefix), self.serve_layout)
-
-        self._add_url("{}_dash-dependencies".format(prefix), self.dependencies)
+        self.server.errorhandler(InvalidResourceError)(self._invalid_resources_handler)
 
         self._add_url(
-            "{}_dash-update-component".format(prefix), self.dispatch, ["POST"]
-        )
-
-        self._add_url(
-            (
-                "{}_dash-component-suites"
-                "/<string:package_name>"
-                "/<path:path_in_package_dist>"
-            ).format(prefix),
+            "_dash-component-suites/<string:package_name>/<path:fingerprinted_path>",
             self.serve_component_suites,
         )
-
-        self._add_url("{}_dash-routes".format(prefix), self.serve_routes)
-
-        self._add_url(prefix, self.index)
-
-        self._add_url("{}_reload-hash".format(prefix), self.serve_reload_hash)
+        self._add_url("_dash-layout", self.serve_layout)
+        self._add_url("_dash-dependencies", self.dependencies)
+        self._add_url("_dash-update-component", self.dispatch, ["POST"])
+        self._add_url("_reload-hash", self.serve_reload_hash)
+        self._add_url("_favicon.ico", self._serve_default_favicon)
+        self._add_url("", self.index)
 
         # catch-all for front-end routes, used by dcc.Location
-        self._add_url("{}<path:path>".format(prefix), self.index)
-
-        self._add_url(
-            "{}_favicon.ico".format(prefix), self._serve_default_favicon
-        )
+        self._add_url("<path:path>", self.index)
 
     def _add_url(self, name, view_func, methods=("GET",)):
+        full_name = self.config.routes_pathname_prefix + name
+
         self.server.add_url_rule(
-            name, view_func=view_func, endpoint=name, methods=list(methods)
+            full_name, view_func=view_func, endpoint=full_name, methods=list(methods)
         )
 
         # record the url in Dash.routes so that it can be accessed later
         # e.g. for adding authentication with flask_login
-        self.routes.append(name)
+        self.routes.append(full_name)
 
     @property
     def layout(self):
         return self._layout
 
     def _layout_value(self):
-        if isinstance(self._layout, _patch_collections_abc("Callable")):
+        if isinstance(self._layout, patch_collections_abc("Callable")):
             self._cached_layout = self._layout()
         else:
             self._cached_layout = self._layout
@@ -446,15 +429,7 @@ class Dash(object):
 
     @layout.setter
     def layout(self, value):
-        if not isinstance(value, Component) and not isinstance(
-            value, _patch_collections_abc("Callable")
-        ):
-            raise exceptions.NoLayoutException(
-                "Layout must be a dash component "
-                "or a function that returns "
-                "a dash component."
-            )
-
+        _validate.validate_layout_type(value)
         self._cached_layout = None
         self._layout = value
 
@@ -464,18 +439,8 @@ class Dash(object):
 
     @index_string.setter
     def index_string(self, value):
-        checks = (
-            (_re_index_entry.search(value), "app_entry"),
-            (_re_index_config.search(value), "config"),
-            (_re_index_scripts.search(value), "scripts"),
-        )
-        missing = [missing for check, missing in checks if not check]
-        if missing:
-            raise exceptions.InvalidIndexException(
-                "Did you forget to include {} in your index string ?".format(
-                    ", ".join("{%" + x + "%}" for x in missing)
-                )
-            )
+        checks = (_re_index_entry, _re_index_config, _re_index_scripts)
+        _validate.validate_index("index string", checks, value)
         self._index_string = value
 
     def serve_layout(self):
@@ -495,6 +460,7 @@ class Dash(object):
             "ui": self._dev_tools.ui,
             "props_check": self._dev_tools.props_check,
             "show_undo_redo": self.config.show_undo_redo,
+            "suppress_callback_exceptions": self.config.suppress_callback_exceptions,
         }
         if self._dev_tools.hot_reload:
             config["hot_reload"] = {
@@ -522,12 +488,6 @@ class Dash(object):
             }
         )
 
-    def serve_routes(self):
-        return flask.Response(
-            json.dumps(self.routes, cls=plotly.utils.PlotlyJSONEncoder),
-            mimetype="application/json",
-        )
-
     def _collect_and_register_resources(self, resources):
         # now needs the app context.
         # template in the necessary component suite JS bundles
@@ -536,8 +496,7 @@ class Dash(object):
         def _relative_url_path(relative_package_path="", namespace=""):
 
             module_path = os.path.join(
-                os.path.dirname(sys.modules[namespace].__file__),
-                relative_package_path,
+                os.path.dirname(sys.modules[namespace].__file__), relative_package_path
             )
 
             modified = int(os.stat(module_path).st_mtime)
@@ -577,9 +536,7 @@ class Dash(object):
                     else:
                         srcs += resource["external_url"]
             elif "absolute_path" in resource:
-                raise Exception(
-                    "Serving files from absolute_path isn't supported yet"
-                )
+                raise Exception("Serving files from absolute_path isn't supported yet")
             elif "asset_path" in resource:
                 static_url = self.get_asset_url(resource["asset_path"])
                 # Add a cache-busting query param
@@ -593,7 +550,7 @@ class Dash(object):
 
         return "\n".join(
             [
-                _format_tag("link", link, opened=True)
+                format_tag("link", link, opened=True)
                 if isinstance(link, dict)
                 else '<link rel="stylesheet" href="{}">'.format(link)
                 for link in (external_links + links)
@@ -622,9 +579,7 @@ class Dash(object):
         dev = self._dev_tools.serve_dev_bundles
         srcs = (
             self._collect_and_register_resources(
-                self.scripts._resources._filter_resources(
-                    deps, dev_bundles=dev
-                )
+                self.scripts._resources._filter_resources(deps, dev_bundles=dev)
             )
             + self.config.external_scripts
             + self._collect_and_register_resources(
@@ -637,23 +592,18 @@ class Dash(object):
 
         return "\n".join(
             [
-                _format_tag("script", src)
+                format_tag("script", src)
                 if isinstance(src, dict)
                 else '<script src="{}"></script>'.format(src)
                 for src in srcs
-            ] +
-            [
-                '<script>{}</script>'.format(src)
-                for src in self._inline_scripts
             ]
+            + ["<script>{}</script>".format(src) for src in self._inline_scripts]
         )
 
     def _generate_config_html(self):
-        return (
-            '<script id="_dash-config" type="application/json">'
-            "{}"
-            "</script>"
-        ).format(json.dumps(self._config()))
+        return ('<script id="_dash-config" type="application/json">{}</script>').format(
+            json.dumps(self._config())
+        )
 
     def _generate_renderer(self):
         return (
@@ -671,39 +621,19 @@ class Dash(object):
 
         tags = []
         if not has_ie_compat:
-            tags.append(
-                '<meta http-equiv="X-UA-Compatible" content="IE=edge">'
-            )
+            tags.append('<meta http-equiv="X-UA-Compatible" content="IE=edge">')
         if not has_charset:
             tags.append('<meta charset="UTF-8">')
 
-        tags += [_format_tag("meta", x, opened=True) for x in meta_tags]
+        tags += [format_tag("meta", x, opened=True) for x in meta_tags]
 
         return "\n      ".join(tags)
 
     # Serve the JS bundles for each package
-    def serve_component_suites(self, package_name, path_in_package_dist):
-        path_in_package_dist, has_fingerprint = check_fingerprint(
-            path_in_package_dist
-        )
+    def serve_component_suites(self, package_name, fingerprinted_path):
+        path_in_pkg, has_fingerprint = check_fingerprint(fingerprinted_path)
 
-        if package_name not in self.registered_paths:
-            raise exceptions.DependencyException(
-                "Error loading dependency.\n"
-                '"{}" is not a registered library.\n'
-                "Registered libraries are: {}".format(
-                    package_name, list(self.registered_paths.keys())
-                )
-            )
-
-        if path_in_package_dist not in self.registered_paths[package_name]:
-            raise exceptions.DependencyException(
-                '"{}" is registered but the path requested is not valid.\n'
-                'The path requested: "{}"\n'
-                "List of registered paths: {}".format(
-                    package_name, path_in_package_dist, self.registered_paths
-                )
-            )
+        _validate.validate_js_path(self.registered_paths, package_name, path_in_pkg)
 
         mimetype = (
             {
@@ -711,20 +641,19 @@ class Dash(object):
                 "css": "text/css",
                 "map": "application/json",
             }
-        )[path_in_package_dist.split(".")[-1]]
+        )[path_in_pkg.split(".")[-1]]
 
         package = sys.modules[package_name]
         self.logger.debug(
             "serving -- package: %s[%s] resource: %s => location: %s",
             package_name,
             package.__version__,
-            path_in_package_dist,
+            path_in_pkg,
             package.__path__,
         )
 
         response = flask.Response(
-            pkgutil.get_data(package_name, path_in_package_dist),
-            mimetype=mimetype,
+            pkgutil.get_data(package_name, path_in_pkg), mimetype=mimetype
         )
 
         if has_fingerprint:
@@ -737,7 +666,7 @@ class Dash(object):
             response.add_etag()
             tag = response.get_etag()[0]
 
-            request_etag = flask.request.headers.get('If-None-Match')
+            request_etag = flask.request.headers.get("If-None-Match")
 
             if '"{}"'.format(tag) == request_etag:
                 response = flask.Response(None, status=304)
@@ -764,7 +693,7 @@ class Dash(object):
                 self.config.requests_pathname_prefix, __version__
             )
 
-        favicon = _format_tag(
+        favicon = format_tag(
             "link",
             {"rel": "icon", "type": "image/x-icon", "href": favicon_url},
             opened=True,
@@ -782,21 +711,12 @@ class Dash(object):
         )
 
         checks = (
-            (_re_index_entry_id.search(index), "#react-entry-point"),
-            (_re_index_config_id.search(index), "#_dash-configs"),
-            (_re_index_scripts_id.search(index), "dash-renderer"),
-            (_re_renderer_scripts_id.search(index), "new DashRenderer"),
+            _re_index_entry_id,
+            _re_index_config_id,
+            _re_index_scripts_id,
+            _re_renderer_scripts_id,
         )
-        missing = [missing for check, missing in checks if not check]
-
-        if missing:
-            plural = "s" if len(missing) > 1 else ""
-            raise exceptions.InvalidIndexException(
-                "Missing element{pl} {ids} in index.".format(
-                    ids=", ".join(missing), pl=plural
-                )
-            )
-
+        _validate.validate_index("index", checks, index)
         return index
 
     def interpolate_index(
@@ -845,7 +765,7 @@ class Dash(object):
         :param favicon: A favicon <link> tag if found in assets folder.
         :return: The interpolated HTML string for the index.
         """
-        return _interpolate(
+        return interpolate_str(
             self.index_string,
             metas=metas,
             title=title,
@@ -858,351 +778,30 @@ class Dash(object):
         )
 
     def dependencies(self):
-        return flask.jsonify(
-            [
-                {
-                    "output": k,
-                    "inputs": v["inputs"],
-                    "state": v["state"],
-                    "clientside_function": v.get("clientside_function", None),
-                }
-                for k, v in self.callback_map.items()
-            ]
-        )
+        return flask.jsonify(self._callback_list)
 
-    def _validate_callback(self, output, inputs, state):
-        # pylint: disable=too-many-branches
-        layout = self._cached_layout or self._layout_value()
-        is_multi = isinstance(output, (list, tuple))
+    def _insert_callback(self, output, inputs, state):
+        _validate.validate_callback(output, inputs, state)
+        callback_id = create_callback_id(output)
+        callback_spec = {
+            "output": callback_id,
+            "inputs": [c.to_dict() for c in inputs],
+            "state": [c.to_dict() for c in state],
+            "clientside_function": None,
+        }
+        self.callback_map[callback_id] = {
+            "inputs": callback_spec["inputs"],
+            "state": callback_spec["state"],
+        }
+        self._callback_list.append(callback_spec)
 
-        if layout is None and not self.config.suppress_callback_exceptions:
-            # Without a layout, we can't do validation on the IDs and
-            # properties of the elements in the callback.
-            raise exceptions.LayoutIsNotDefined(
-                dedent(
-                    """
-                Attempting to assign a callback to the application but
-                the `layout` property has not been assigned.
-                Assign the `layout` property before assigning callbacks.
-                Alternatively, suppress this warning by setting
-                `suppress_callback_exceptions=True`
-            """
-                )
-            )
+        return callback_id
 
-        outputs = output if is_multi else [output]
-        for args, obj, name in [
-            (outputs, Output, "Output"),
-            (inputs, Input, "Input"),
-            (state, State, "State"),
-        ]:
-
-            if not isinstance(args, (list, tuple)):
-                raise exceptions.IncorrectTypeException(
-                    "The {} argument `{}` must be "
-                    "a list or tuple of `dash.dependencies.{}`s.".format(
-                        name.lower(), str(args), name
-                    )
-                )
-
-            for arg in args:
-                if not isinstance(arg, obj):
-                    raise exceptions.IncorrectTypeException(
-                        "The {} argument `{}` must be "
-                        "of type `dash.{}`.".format(
-                            name.lower(), str(arg), name
-                        )
-                    )
-
-                invalid_characters = ["."]
-                if any(x in arg.component_id for x in invalid_characters):
-                    raise exceptions.InvalidComponentIdError(
-                        "The element `{}` contains {} in its ID. "
-                        "Periods are not allowed in IDs.".format(
-                            arg.component_id, invalid_characters
-                        )
-                    )
-
-                if not self.config.suppress_callback_exceptions:
-                    layout_id = getattr(layout, "id", None)
-                    arg_id = arg.component_id
-                    arg_prop = getattr(arg, "component_property", None)
-                    if arg_id not in layout and arg_id != layout_id:
-                        all_ids = [k for k in layout]
-                        if layout_id:
-                            all_ids.append(layout_id)
-                        raise exceptions.NonExistentIdException(
-                            dedent(
-                                """
-                            Attempting to assign a callback to the
-                            component with the id "{0}" but no
-                            components with id "{0}" exist in the
-                            app\'s layout.\n\n
-                            Here is a list of IDs in layout:\n{1}\n\n
-                            If you are assigning callbacks to components
-                            that are generated by other callbacks
-                            (and therefore not in the initial layout), then
-                            you can suppress this exception by setting
-                            `suppress_callback_exceptions=True`.
-                        """
-                            ).format(arg_id, all_ids)
-                        )
-
-                    component = (
-                        layout if layout_id == arg_id else layout[arg_id]
-                    )
-
-                    if (
-                        arg_prop
-                        and arg_prop not in component.available_properties
-                        and not any(
-                            arg_prop.startswith(w)
-                            for w in component.available_wildcard_properties
-                        )
-                    ):
-                        raise exceptions.NonExistentPropException(
-                            dedent(
-                                """
-                            Attempting to assign a callback with
-                            the property "{0}" but the component
-                            "{1}" doesn't have "{0}" as a property.\n
-                            Here are the available properties in "{1}":
-                            {2}
-                        """
-                            ).format(
-                                arg_prop,
-                                arg_id,
-                                component.available_properties,
-                            )
-                        )
-
-                    if hasattr(arg, "component_event"):
-                        raise exceptions.NonExistentEventException(
-                            dedent(
-                                """
-                            Events have been removed.
-                            Use the associated property instead.
-                        """
-                            )
-                        )
-
-        if state and not inputs:
-            raise exceptions.MissingInputsException(
-                dedent(
-                    """
-                This callback has {} `State` {}
-                but no `Input` elements.\n
-                Without `Input` elements, this callback
-                will never get called.\n
-                (Subscribing to input components will cause the
-                callback to be called whenever their values change.)
-            """
-                ).format(
-                    len(state), "elements" if len(state) > 1 else "element"
-                )
-            )
-
-        for i in inputs:
-            bad = None
-            if is_multi:
-                for o in output:
-                    if o == i:
-                        bad = o
-            else:
-                if output == i:
-                    bad = output
-            if bad:
-                raise exceptions.SameInputOutputException(
-                    "Same output and input: {}".format(bad)
-                )
-
-        if is_multi:
-            if len(set(output)) != len(output):
-                raise exceptions.DuplicateCallbackOutput(
-                    "Same output was used more than once in a "
-                    "multi output callback!\n Duplicates:\n {}".format(
-                        ",\n".join(
-                            k
-                            for k, v in (
-                                (str(x), output.count(x)) for x in output
-                            )
-                            if v > 1
-                        )
-                    )
-                )
-
-        callback_id = _create_callback_id(output)
-
-        callbacks = set(
-            itertools.chain(
-                *(
-                    x[2:-2].split("...") if x.startswith("..") else [x]
-                    for x in self.callback_map
-                )
-            )
-        )
-        ns = {"duplicates": set()}
-        if is_multi:
-
-            def duplicate_check():
-                ns["duplicates"] = callbacks.intersection(
-                    str(y) for y in output
-                )
-                return ns["duplicates"]
-
-        else:
-
-            def duplicate_check():
-                return callback_id in callbacks
-
-        if duplicate_check():
-            if is_multi:
-                msg = dedent(
-                    """
-                Multi output {} contains an `Output` object
-                that was already assigned.
-                Duplicates:
-                {}
-                """
-                ).format(callback_id, pprint.pformat(ns["duplicates"]))
-            else:
-                msg = dedent(
-                    """
-                You have already assigned a callback to the output
-                with ID "{}" and property "{}". An output can only have
-                a single callback function. Try combining your inputs and
-                callback functions together into one function.
-                """
-                ).format(output.component_id, output.component_property)
-            raise exceptions.DuplicateCallbackOutput(msg)
-
-    @staticmethod
-    def _validate_callback_output(output_value, output):
-        valid = [str, dict, int, float, type(None), Component]
-
-        def _raise_invalid(
-            bad_val, outer_val, path, index=None, toplevel=False
-        ):
-            bad_type = type(bad_val).__name__
-            outer_id = (
-                "(id={:s})".format(outer_val.id)
-                if getattr(outer_val, "id", False)
-                else ""
-            )
-            outer_type = type(outer_val).__name__
-            raise exceptions.InvalidCallbackReturnValue(
-                dedent(
-                    """
-            The callback for `{output:s}`
-            returned a {object:s} having type `{type:s}`
-            which is not JSON serializable.
-
-            {location_header:s}{location:s}
-            and has string representation
-            `{bad_val}`
-
-            In general, Dash properties can only be
-            dash components, strings, dictionaries, numbers, None,
-            or lists of those.
-            """
-                ).format(
-                    output=repr(output),
-                    object="tree with one value" if not toplevel else "value",
-                    type=bad_type,
-                    location_header=(
-                        "The value in question is located at"
-                        if not toplevel
-                        else "The value in question is either the only value "
-                        "returned,\nor is in the top level of the returned "
-                        "list,"
-                    ),
-                    location=(
-                        "\n"
-                        + (
-                            "[{:d}] {:s} {:s}".format(
-                                index, outer_type, outer_id
-                            )
-                            if index is not None
-                            else ("[*] " + outer_type + " " + outer_id)
-                        )
-                        + "\n"
-                        + path
-                        + "\n"
-                    )
-                    if not toplevel
-                    else "",
-                    bad_val=bad_val,
-                )
-            )
-
-        def _value_is_valid(val):
-            return (
-                # pylint: disable=unused-variable
-                any([isinstance(val, x) for x in valid])
-                or type(val).__name__ == "unicode"
-            )
-
-        def _validate_value(val, index=None):
-            # val is a Component
-            if isinstance(val, Component):
-                # pylint: disable=protected-access
-                for p, j in val._traverse_with_paths():
-                    # check each component value in the tree
-                    if not _value_is_valid(j):
-                        _raise_invalid(
-                            bad_val=j, outer_val=val, path=p, index=index
-                        )
-
-                    # Children that are not of type Component or
-                    # list/tuple not returned by traverse
-                    child = getattr(j, "children", None)
-                    if not isinstance(
-                        child, (tuple, collections.MutableSequence)
-                    ):
-                        if child and not _value_is_valid(child):
-                            _raise_invalid(
-                                bad_val=child,
-                                outer_val=val,
-                                path=p + "\n" + "[*] " + type(child).__name__,
-                                index=index,
-                            )
-
-                # Also check the child of val, as it will not be returned
-                child = getattr(val, "children", None)
-                if not isinstance(child, (tuple, collections.MutableSequence)):
-                    if child and not _value_is_valid(child):
-                        _raise_invalid(
-                            bad_val=child,
-                            outer_val=val,
-                            path=type(child).__name__,
-                            index=index,
-                        )
-
-            # val is not a Component, but is at the top level of tree
-            else:
-                if not _value_is_valid(val):
-                    _raise_invalid(
-                        bad_val=val,
-                        outer_val=type(val).__name__,
-                        path="",
-                        index=index,
-                        toplevel=True,
-                    )
-
-        if isinstance(output_value, list):
-            for i, val in enumerate(output_value):
-                _validate_value(val, index=i)
-        else:
-            _validate_value(output_value)
-
-    # pylint: disable=dangerous-default-value
-    def clientside_callback(
-        self, clientside_function, output, inputs=[], state=[]
-    ):
+    def clientside_callback(self, clientside_function, output, inputs, state=()):
         """Create a callback that updates the output by calling a clientside
         (JavaScript) function instead of a Python function.
 
-        Unlike `@app.calllback`, `clientside_callback` is not a decorator:
+        Unlike `@app.callback`, `clientside_callback` is not a decorator:
         it takes either a
         `dash.dependencies.ClientsideFunction(namespace, function_name)`
         argument that describes which JavaScript function to call
@@ -1259,8 +858,7 @@ class Dash(object):
         )
         ```
         """
-        self._validate_callback(output, inputs, state)
-        callback_id = _create_callback_id(output)
+        self._insert_callback(output, inputs, state)
 
         # If JS source is explicitly given, create a namespace and function
         # name, then inject the code.
@@ -1270,17 +868,15 @@ class Dash(object):
             if isinstance(output, (list, tuple)):
                 out0 = output[0]
 
-            namespace = '_dashprivate_{}'.format(out0.component_id)
-            function_name = '{}'.format(out0.component_property)
+            namespace = "_dashprivate_{}".format(out0.component_id)
+            function_name = "{}".format(out0.component_property)
 
             self._inline_scripts.append(
-                """
-                var clientside = window.dash_clientside = window.dash_clientside || {{}};
-                var ns = clientside["{0}"] = clientside["{0}"] || {{}};
-                ns["{1}"] = {2};
-                """.format(namespace.replace('"', '\\"'),
-                           function_name.replace('"', '\\"'),
-                           clientside_function)
+                _inline_clientside_template.format(
+                    namespace=namespace.replace('"', '\\"'),
+                    function_name=function_name.replace('"', '\\"'),
+                    clientside_function=clientside_function,
+                )
             )
 
         # Callback is stored in an external asset.
@@ -1288,118 +884,57 @@ class Dash(object):
             namespace = clientside_function.namespace
             function_name = clientside_function.function_name
 
-        self.callback_map[callback_id] = {
-            "inputs": [
-                {"id": c.component_id, "property": c.component_property}
-                for c in inputs
-            ],
-            "state": [
-                {"id": c.component_id, "property": c.component_property}
-                for c in state
-            ],
-            "clientside_function": {
-                "namespace": namespace,
-                "function_name": function_name,
-            },
+        self._callback_list[-1]["clientside_function"] = {
+            "namespace": namespace,
+            "function_name": function_name,
         }
 
-    # TODO - Update nomenclature.
-    # "Parents" and "Children" should refer to the DOM tree
-    # and not the dependency tree.
-    # The dependency tree should use the nomenclature
-    # "observer" and "controller".
-    # "observers" listen for changes from their "controllers". For example,
-    # if a graph depends on a dropdown, the graph is the "observer" and the
-    # dropdown is a "controller". In this case the graph's "dependency" is
-    # the dropdown.
-    # TODO - Check this map for recursive or other ill-defined non-tree
-    # relationships
-    # pylint: disable=dangerous-default-value
-    def callback(self, output, inputs=[], state=[]):
-        self._validate_callback(output, inputs, state)
-
-        callback_id = _create_callback_id(output)
+    def callback(self, output, inputs, state=()):
+        callback_id = self._insert_callback(output, inputs, state)
         multi = isinstance(output, (list, tuple))
-
-        self.callback_map[callback_id] = {
-            "inputs": [
-                {"id": c.component_id, "property": c.component_property}
-                for c in inputs
-            ],
-            "state": [
-                {"id": c.component_id, "property": c.component_property}
-                for c in state
-            ],
-        }
 
         def wrap_func(func):
             @wraps(func)
             def add_context(*args, **kwargs):
+                output_spec = kwargs.pop("outputs_list")
+
                 # don't touch the comment on the next line - used by debugger
                 output_value = func(*args, **kwargs)  # %% callback invoked %%
-                if multi:
-                    if not isinstance(output_value, (list, tuple)):
-                        raise exceptions.InvalidCallbackReturnValue(
-                            "The callback {} is a multi-output.\n"
-                            "Expected the output type to be a list"
-                            " or tuple but got {}.".format(
-                                callback_id, repr(output_value)
-                            )
-                        )
 
-                    if not len(output_value) == len(output):
-                        raise exceptions.InvalidCallbackReturnValue(
-                            "Invalid number of output values for {}.\n"
-                            " Expected {} got {}".format(
-                                callback_id, len(output), len(output_value)
-                            )
-                        )
+                if isinstance(output_value, _NoUpdate):
+                    raise PreventUpdate
 
-                    component_ids = collections.defaultdict(dict)
-                    has_update = False
-                    for i, o in enumerate(output):
-                        val = output_value[i]
-                        if not isinstance(val, _NoUpdate):
+                # wrap single outputs so we can treat them all the same
+                # for validation and response creation
+                if not multi:
+                    output_value, output_spec = [output_value], [output_spec]
+
+                _validate.validate_multi_return(output_spec, output_value, callback_id)
+
+                component_ids = collections.defaultdict(dict)
+                has_update = False
+                for val, spec in zip(output_value, output_spec):
+                    if isinstance(val, _NoUpdate):
+                        continue
+                    for vali, speci in (
+                        zip(val, spec) if isinstance(spec, list) else [[val, spec]]
+                    ):
+                        if not isinstance(vali, _NoUpdate):
                             has_update = True
-                            o_id, o_prop = o.component_id, o.component_property
-                            component_ids[o_id][o_prop] = val
+                            id_str = stringify_id(speci["id"])
+                            component_ids[id_str][speci["property"]] = vali
 
-                    if not has_update:
-                        raise exceptions.PreventUpdate
+                if not has_update:
+                    raise PreventUpdate
 
-                    response = {"response": component_ids, "multi": True}
-                else:
-                    if isinstance(output_value, _NoUpdate):
-                        raise exceptions.PreventUpdate
-
-                    response = {
-                        "response": {
-                            "props": {output.component_property: output_value}
-                        }
-                    }
+                response = {"response": component_ids, "multi": True}
 
                 try:
                     jsonResponse = json.dumps(
                         response, cls=plotly.utils.PlotlyJSONEncoder
                     )
                 except TypeError:
-                    self._validate_callback_output(output_value, output)
-                    raise exceptions.InvalidCallbackReturnValue(
-                        dedent(
-                            """
-                    The callback for property `{property:s}`
-                    of component `{id:s}` returned a value
-                    which is not JSON serializable.
-
-                    In general, Dash properties can only be
-                    dash components, strings, dictionaries, numbers, None,
-                    or lists of those.
-                    """
-                        ).format(
-                            property=output.component_property,
-                            id=output.component_id,
-                        )
-                    )
+                    _validate.fail_callback_output(output_value, output)
 
                 return jsonResponse
 
@@ -1411,84 +946,33 @@ class Dash(object):
 
     def dispatch(self):
         body = flask.request.get_json()
-        inputs = body.get("inputs", [])
-        state = body.get("state", [])
+        flask.g.inputs_list = inputs = body.get("inputs", [])
+        flask.g.states_list = state = body.get("state", [])
         output = body["output"]
+        outputs_list = body.get("outputs") or split_callback_id(output)
+        flask.g.outputs_list = outputs_list
 
-        args = []
+        flask.g.input_values = input_values = inputs_to_dict(inputs)
+        flask.g.state_values = inputs_to_dict(state)
+        changed_props = body.get("changedPropIds", [])
+        flask.g.triggered_inputs = [
+            {"prop_id": x, "value": input_values.get(x)} for x in changed_props
+        ]
 
-        flask.g.input_values = input_values = {
-            "{}.{}".format(x["id"], x["property"]): x.get("value")
-            for x in inputs
-        }
-        flask.g.state_values = {
-            "{}.{}".format(x["id"], x["property"]): x.get("value")
-            for x in state
-        }
-        changed_props = body.get("changedPropIds")
-        flask.g.triggered_inputs = (
-            [{"prop_id": x, "value": input_values[x]} for x in changed_props]
-            if changed_props
-            else []
-        )
+        response = flask.g.dash_response = flask.Response(mimetype="application/json")
 
-        response = flask.g.dash_response = flask.Response(
-            mimetype="application/json"
-        )
+        args = inputs_to_vals(inputs) + inputs_to_vals(state)
 
-        for component_registration in self.callback_map[output]["inputs"]:
-            args.append(
-                [
-                    c.get("value", None)
-                    for c in inputs
-                    if c["property"] == component_registration["property"]
-                    and c["id"] == component_registration["id"]
-                ][0]
-            )
-
-        for component_registration in self.callback_map[output]["state"]:
-            args.append(
-                [
-                    c.get("value", None)
-                    for c in state
-                    if c["property"] == component_registration["property"]
-                    and c["id"] == component_registration["id"]
-                ][0]
-            )
-
-        response.set_data(self.callback_map[output]["callback"](*args))
+        func = self.callback_map[output]["callback"]
+        response.set_data(func(*args, outputs_list=outputs_list))
         return response
-
-    def _validate_layout(self):
-        if self.layout is None:
-            raise exceptions.NoLayoutException(
-                "The layout was `None` "
-                "at the time that `run_server` was called. "
-                "Make sure to set the `layout` attribute of your application "
-                "before running the server."
-            )
-
-        to_validate = self._layout_value()
-
-        layout_id = getattr(self.layout, "id", None)
-
-        component_ids = {layout_id} if layout_id else set()
-        # pylint: disable=protected-access
-        for component in to_validate._traverse():
-            component_id = getattr(component, "id", None)
-            if component_id and component_id in component_ids:
-                raise exceptions.DuplicateIdError(
-                    "Duplicate component id found"
-                    " in the initial layout: `{}`".format(component_id)
-                )
-            component_ids.add(component_id)
 
     def _setup_server(self):
         # Apply _force_eager_loading overrides from modules
         eager_loading = self.config.eager_loading
         for module_name in ComponentRegistry.registry:
             module = sys.modules[module_name]
-            eager = getattr(module, '_force_eager_loading', False)
+            eager = getattr(module, "_force_eager_loading", False)
             eager_loading = eager_loading or eager
 
         # Update eager_loading settings
@@ -1497,7 +981,7 @@ class Dash(object):
         if self.config.include_assets_files:
             self._walk_assets_directory()
 
-        self._validate_layout()
+        _validate.validate_layout(self.layout, self._layout_value())
 
         self._generate_scripts_html()
         self._generate_css_dist_html()
@@ -1539,9 +1023,7 @@ class Dash(object):
                 full = os.path.join(current, f)
 
                 if f.endswith("js"):
-                    self.scripts.append_script(
-                        self._add_assets_resource(path, full)
-                    )
+                    self.scripts.append_script(self._add_assets_resource(path, full))
                 elif f.endswith("css"):
                     self.css.append_css(self._add_assets_resource(path, full))
                 elif f == "favicon.ico":
@@ -1554,12 +1036,11 @@ class Dash(object):
     @staticmethod
     def _serve_default_favicon():
         return flask.Response(
-            pkgutil.get_data("dash", "favicon.ico"),
-            content_type="image/x-icon",
+            pkgutil.get_data("dash", "favicon.ico"), content_type="image/x-icon"
         )
 
     def get_asset_url(self, path):
-        asset = _get_asset_path(
+        asset = get_asset_path(
             self.config.requests_pathname_prefix,
             path,
             self.config.assets_url_path.lstrip("/"),
@@ -1604,10 +1085,7 @@ class Dash(object):
                 return chapters.page_2
         ```
         """
-        asset = _get_relative_path(
-            self.config.requests_pathname_prefix,
-            path,
-        )
+        asset = get_relative_path(self.config.requests_pathname_prefix, path)
 
         return asset
 
@@ -1658,14 +1136,11 @@ class Dash(object):
         `page-1/sub-page-1`
         ```
         """
-        return _strip_relative_path(
-            self.config.requests_pathname_prefix,
-            path,
-        )
+        return strip_relative_path(self.config.requests_pathname_prefix, path)
 
     def _setup_dev_tools(self, **kwargs):
         debug = kwargs.get("debug", False)
-        dev_tools = self._dev_tools = _AttributeDict()
+        dev_tools = self._dev_tools = AttributeDict()
 
         for attr in (
             "ui",
@@ -1685,9 +1160,7 @@ class Dash(object):
             ("hot_reload_max_retry", int, 8),
         ):
             dev_tools[attr] = _type(
-                get_combined_config(
-                    attr, kwargs.get(attr, None), default=default
-                )
+                get_combined_config(attr, kwargs.get(attr, None), default=default)
             )
 
         return dev_tools
@@ -1798,7 +1271,7 @@ class Dash(object):
 
         if dev_tools.hot_reload:
             _reload = self._hot_reload
-            _reload.hash = _generate_hash()
+            _reload.hash = generate_hash()
 
             component_packages_dist = [
                 os.path.dirname(package.path)
@@ -1806,8 +1279,7 @@ class Dash(object):
                 else package.filename
                 for package in (
                     pkgutil.find_loader(x)
-                    for x in list(ComponentRegistry.registry)
-                    + ["dash_renderer"]
+                    for x in list(ComponentRegistry.registry) + ["dash_renderer"]
                 )
             ]
 
@@ -1856,15 +1328,13 @@ class Dash(object):
         _reload = self._hot_reload
         with _reload.lock:
             _reload.hard = True
-            _reload.hash = _generate_hash()
+            _reload.hash = generate_hash()
 
             if self.config.assets_folder in filename:
                 asset_path = (
                     os.path.relpath(
                         filename,
-                        os.path.commonprefix(
-                            [self.config.assets_folder, filename]
-                        ),
+                        os.path.commonprefix([self.config.assets_folder, filename]),
                     )
                     .replace("\\", "/")
                     .lstrip("/")
@@ -1907,7 +1377,8 @@ class Dash(object):
 
     def run_server(
         self,
-        port=8050,
+        host=os.getenv("HOST", "127.0.0.1"),
+        port=os.getenv("PORT", "8050"),
         debug=False,
         dev_tools_ui=None,
         dev_tools_props_check=None,
@@ -1926,7 +1397,12 @@ class Dash(object):
         If a parameter can be set by an environment variable, that is listed
         too. Values provided here take precedence over environment variables.
 
+        :param host: Host IP used to serve the application
+            env: ``HOST``
+        :type host: string
+
         :param port: Port used to serve the application
+            env: ``PORT``
         :type port: int
 
         :param debug: Set Flask debug mode and enable dev tools.
@@ -1999,9 +1475,18 @@ class Dash(object):
             dev_tools_prune_errors,
         )
 
+        # Verify port value
+        try:
+            port = int(port)
+            assert port in range(1, 65536)
+        except Exception as e:
+            e.args = [
+                "Expecting an integer from 1 to 65535, found port={}".format(repr(port))
+            ]
+            raise
+
         if self._dev_tools.silence_routes_logging:
             # Since it's silenced, the address doesn't show anymore.
-            host = flask_run_options.get("host", "127.0.0.1")
             ssl_context = flask_run_options.get("ssl_context")
             self.logger.info(
                 "Running on %s://%s:%s%s",
@@ -2021,4 +1506,4 @@ class Dash(object):
 
             self.logger.info("Debugger PIN: %s", debugger_pin)
 
-        self.server.run(port=port, debug=debug, **flask_run_options)
+        self.server.run(host=host, port=port, debug=debug, **flask_run_options)
