@@ -4,11 +4,7 @@ from .consts import SELENIUM_GRID_DEFAULT
 
 
 try:
-    from dash.testing.application_runners import (
-        ThreadedRunner,
-        ProcessRunner,
-        RRunner,
-    )
+    from dash.testing.application_runners import ThreadedRunner, ProcessRunner, RRunner
     from dash.testing.browser import Browser
     from dash.testing.composite import DashComposite, DashRComposite
 except ImportError:
@@ -26,9 +22,7 @@ def pytest_addoption(parser):
     )
 
     dash.addoption(
-        "--remote",
-        action="store_true",
-        help="instruct pytest to use selenium grid",
+        "--remote", action="store_true", help="instruct pytest to use selenium grid"
     )
 
     dash.addoption(
@@ -39,9 +33,7 @@ def pytest_addoption(parser):
     )
 
     dash.addoption(
-        "--headless",
-        action="store_true",
-        help="set this flag to run in headless mode",
+        "--headless", action="store_true", help="set this flag to run in headless mode"
     )
 
     dash.addoption(
@@ -55,6 +47,12 @@ def pytest_addoption(parser):
         "--nopercyfinalize",
         action="store_false",
         help="set this flag to control percy finalize at CI level",
+    )
+
+    dash.addoption(
+        "--pause",
+        action="store_true",
+        help="pause using pdb after opening the test app, so you can interact with it",
     )
 
 
@@ -122,6 +120,7 @@ def dash_br(request, tmpdir):
         download_path=tmpdir.mkdir("download").strpath,
         percy_assets_root=request.config.getoption("percy_assets"),
         percy_finalize=request.config.getoption("nopercyfinalize"),
+        pause=request.config.getoption("pause"),
     ) as browser:
         yield browser
 
@@ -138,6 +137,7 @@ def dash_duo(request, dash_thread_server, tmpdir):
         download_path=tmpdir.mkdir("download").strpath,
         percy_assets_root=request.config.getoption("percy_assets"),
         percy_finalize=request.config.getoption("nopercyfinalize"),
+        pause=request.config.getoption("pause"),
     ) as dc:
         yield dc
 
@@ -154,5 +154,6 @@ def dashr(request, dashr_server, tmpdir):
         download_path=tmpdir.mkdir("download").strpath,
         percy_assets_root=request.config.getoption("percy_assets"),
         percy_finalize=request.config.getoption("nopercyfinalize"),
+        pause=request.config.getoption("pause"),
     ) as dc:
         yield dc
