@@ -10,10 +10,7 @@ logger = logging.getLogger(__name__)
 
 
 def until(
-    wait_cond,
-    timeout,
-    poll=0.1,
-    msg="expected condition not met within timeout",
+    wait_cond, timeout, poll=0.1, msg="expected condition not met within timeout"
 ):  # noqa: C0330
     res = wait_cond()
     logger.debug(
@@ -34,10 +31,7 @@ def until(
 
 
 def until_not(
-    wait_cond,
-    timeout,
-    poll=0.1,
-    msg="expected condition met within timeout",
+    wait_cond, timeout, poll=0.1, msg="expected condition met within timeout"
 ):  # noqa: C0330
     res = wait_cond()
     logger.debug(
@@ -65,14 +59,11 @@ class contains_text(object):
     def __call__(self, driver):
         try:
             elem = driver.find_element_by_css_selector(self.selector)
-            logger.debug(
-                "contains text {%s} => expected %s", elem.text, self.text
-            )
+            logger.debug("contains text {%s} => expected %s", elem.text, self.text)
             return self.text in str(elem.text) or self.text in str(
                 elem.get_attribute("value")
             )
         except WebDriverException:
-            logger.exception("contains_text encountered an exception")
             return False
 
 
@@ -84,15 +75,12 @@ class text_to_equal(object):
     def __call__(self, driver):
         try:
             elem = driver.find_element_by_css_selector(self.selector)
-            logger.debug(
-                "text to equal {%s} => expected %s", elem.text, self.text
-            )
+            logger.debug("text to equal {%s} => expected %s", elem.text, self.text)
             return (
                 str(elem.text) == self.text
                 or str(elem.get_attribute("value")) == self.text
             )
         except WebDriverException:
-            logger.exception("text_to_equal encountered an exception")
             return False
 
 
@@ -109,5 +97,4 @@ class style_to_equal(object):
             logger.debug("style to equal {%s} => expected %s", val, self.val)
             return val == self.val
         except WebDriverException:
-            logger.exception("style_to_equal encountered an exception")
             return False
