@@ -1,1 +1,1 @@
-__version__ = '1.8.0'
+__version__ = "1.10.0"
