@@ -22,10 +22,7 @@ test_cases = {
         "fail": True,
         "name": "invalid nested prop",
         "component": dcc.Checklist,
-        "props": {
-            "options": [{"label": "hello", "value": True}],
-            "value": ["test"],
-        },
+        "props": {"options": [{"label": "hello", "value": True}], "value": ["test"]},
     },
     "invalid-arrayOf": {
         "fail": True,
@@ -56,13 +53,7 @@ test_cases = {
         "name": "nested object with bad value",
         "component": DataTable,
         "props": {
-            "columns": [{
-                "id": "id",
-                "name": "name",
-                "format": {
-                    "locale": "asdf"
-                }
-            }]
+            "columns": [{"id": "id", "name": "name", "format": {"locale": "asdf"}}]
         },
     },
     "invalid-shape-3": {
@@ -70,13 +61,7 @@ test_cases = {
         "name": "invalid oneOf within nested object",
         "component": DataTable,
         "props": {
-            "columns": [{
-                "id": "id",
-                "name": "name",
-                "on_change": {
-                    "action": "asdf"
-                }
-            }]
+            "columns": [{"id": "id", "name": "name", "on_change": {"action": "asdf"}}]
         },
     },
     "invalid-shape-4": {
@@ -84,31 +69,20 @@ test_cases = {
         "name": "invalid key within deeply nested object",
         "component": DataTable,
         "props": {
-            "columns": [{
-                "id": "id",
-                "name": "name",
-                "on_change": {
-                    "asdf": "asdf"
-                }
-            }]
+            "columns": [{"id": "id", "name": "name", "on_change": {"asdf": "asdf"}}]
         },
     },
     "invalid-shape-5": {
         "fail": True,
         "name": "invalid not required key",
         "component": dcc.Dropdown,
-        "props": {
-            "options": [{"label": "new york", "value": "ny", "typo": "asdf"}]
-        },
+        "props": {"options": [{"label": "new york", "value": "ny", "typo": "asdf"}]},
     },
     "string-not-list": {
         "fail": True,
         "name": "string-not-a-list",
         "component": dcc.Checklist,
-        "props": {
-            "options": [{"label": "hello", "value": "test"}],
-            "value": "test",
-        },
+        "props": {"options": [{"label": "hello", "value": "test"}], "value": "test"},
     },
     "no-properties": {
         "fail": False,
@@ -139,13 +113,7 @@ test_cases = {
         "name": "nested string instead of number/null",
         "component": DataTable,
         "props": {
-            "columns": [{
-                "id": "id",
-                "name": "name",
-                "format": {
-                    "prefix": "asdf"
-                }
-            }]
+            "columns": [{"id": "id", "name": "name", "format": {"prefix": "asdf"}}]
         },
     },
     "allow-null": {
@@ -153,13 +121,7 @@ test_cases = {
         "name": "nested null",
         "component": DataTable,
         "props": {
-            "columns": [{
-                "id": "id",
-                "name": "name",
-                "format": {
-                    "prefix": None
-                }
-            }]
+            "columns": [{"id": "id", "name": "name", "format": {"prefix": None}}]
         },
     },
     "allow-null-2": {
@@ -202,9 +164,7 @@ test_cases = {
         "fail": False,
         "name": "allow exact with optional and required keys",
         "component": dcc.Dropdown,
-        "props": {
-            "options": [{"label": "new york", "value": "ny", "disabled": False}]
-        },
+        "props": {"options": [{"label": "new york", "value": "ny", "disabled": False}]},
     },
     "allow-exact-with-optional-and-required-2": {
         "fail": False,
@@ -220,16 +180,13 @@ def test_dvpc001_prop_check_errors_with_path(dash_duo):
 
     app.layout = html.Div([html.Div(id="content"), dcc.Location(id="location")])
 
-    @app.callback(
-        Output("content", "children"), [Input("location", "pathname")]
-    )
+    @app.callback(Output("content", "children"), [Input("location", "pathname")])
     def display_content(pathname):
         if pathname is None or pathname == "/":
             return "Initial state"
         test_case = test_cases[pathname.strip("/")]
         return html.Div(
-            id="new-component",
-            children=test_case["component"](**test_case["props"]),
+            id="new-component", children=test_case["component"](**test_case["props"])
         )
 
     dash_duo.start_server(
@@ -247,14 +204,10 @@ def test_dvpc001_prop_check_errors_with_path(dash_duo):
         if test_cases[tc]["fail"]:
             dash_duo.wait_for_element(".test-devtools-error-toggle").click()
             dash_duo.percy_snapshot(
-                "devtools validation exception: {}".format(
-                    test_cases[tc]["name"]
-                )
+                "devtools validation exception: {}".format(test_cases[tc]["name"])
             )
         else:
             dash_duo.wait_for_element("#new-component")
             dash_duo.percy_snapshot(
-                "devtools validation no exception: {}".format(
-                    test_cases[tc]["name"]
-                )
+                "devtools validation no exception: {}".format(test_cases[tc]["name"])
             )
