@@ -12,11 +12,7 @@ from dash._configs import (
     get_combined_config,
     load_dash_env_vars,
 )
-from dash._utils import (
-    get_asset_path,
-    get_relative_path,
-    strip_relative_path,
-)
+from dash._utils import get_asset_path, get_relative_path, strip_relative_path
 
 
 @pytest.fixture
@@ -167,60 +163,53 @@ def test_app_name_server(empty_environ, name, server, expected):
     [
         ("/", "", "/"),
         ("/my-dash-app/", "", "/my-dash-app/"),
-
         ("/", "/", "/"),
         ("/my-dash-app/", "/", "/my-dash-app/"),
-
         ("/", "/page-1", "/page-1"),
         ("/my-dash-app/", "/page-1", "/my-dash-app/page-1"),
-
         ("/", "/page-1/", "/page-1/"),
         ("/my-dash-app/", "/page-1/", "/my-dash-app/page-1/"),
-
         ("/", "/page-1/sub-page-1", "/page-1/sub-page-1"),
         ("/my-dash-app/", "/page-1/sub-page-1", "/my-dash-app/page-1/sub-page-1"),
-    ]
+    ],
 )
 def test_pathname_prefix_relative_url(prefix, partial_path, expected):
     path = get_relative_path(prefix, partial_path)
     assert path == expected
 
+
 @pytest.mark.parametrize(
     "prefix, partial_path",
-    [
-        ("/", "relative-page-1"),
-        ("/my-dash-app/", "relative-page-1"),
-    ]
+    [("/", "relative-page-1"), ("/my-dash-app/", "relative-page-1")],
 )
 def test_invalid_get_relative_path(prefix, partial_path):
     with pytest.raises(_exc.UnsupportedRelativePath):
         get_relative_path(prefix, partial_path)
+
 
 @pytest.mark.parametrize(
     "prefix, partial_path, expected",
     [
         ("/", None, None),
         ("/my-dash-app/", None, None),
-
         ("/", "/", ""),
         ("/my-dash-app/", "/my-dash-app", ""),
         ("/my-dash-app/", "/my-dash-app/", ""),
-
         ("/", "/page-1", "page-1"),
         ("/my-dash-app/", "/my-dash-app/page-1", "page-1"),
-
         ("/", "/page-1/", "page-1"),
         ("/my-dash-app/", "/my-dash-app/page-1/", "page-1"),
-
         ("/", "/page-1/sub-page-1", "page-1/sub-page-1"),
         ("/my-dash-app/", "/my-dash-app/page-1/sub-page-1", "page-1/sub-page-1"),
-
         ("/", "/page-1/sub-page-1/", "page-1/sub-page-1"),
         ("/my-dash-app/", "/my-dash-app/page-1/sub-page-1/", "page-1/sub-page-1"),
-
         ("/my-dash-app/", "/my-dash-app/my-dash-app/", "my-dash-app"),
-        ("/my-dash-app/", "/my-dash-app/something-else/my-dash-app/", "something-else/my-dash-app"),
-    ]
+        (
+            "/my-dash-app/",
+            "/my-dash-app/something-else/my-dash-app/",
+            "something-else/my-dash-app",
+        ),
+    ],
 )
 def test_strip_relative_path(prefix, partial_path, expected):
     path = strip_relative_path(prefix, partial_path)
@@ -232,9 +221,36 @@ def test_strip_relative_path(prefix, partial_path, expected):
     [
         ("/", "relative-page-1"),
         ("/my-dash-app", "relative-page-1"),
-        ("/my-dash-app", "/some-other-path")
-    ]
+        ("/my-dash-app", "/some-other-path"),
+    ],
 )
 def test_invalid_strip_relative_path(prefix, partial_path):
     with pytest.raises(_exc.UnsupportedRelativePath):
         strip_relative_path(prefix, partial_path)
+
+
+def test_port_env_fail_str(empty_environ):
+    app = Dash()
+    with pytest.raises(Exception) as excinfo:
+        app.run_server(port="garbage")
+    assert (
+        excinfo.exconly()
+        == "ValueError: Expecting an integer from 1 to 65535, found port='garbage'"
+    )
+
+
+def test_port_env_fail_range(empty_environ):
+    app = Dash()
+    with pytest.raises(Exception) as excinfo:
+        app.run_server(port="0")
+    assert (
+        excinfo.exconly()
+        == "AssertionError: Expecting an integer from 1 to 65535, found port=0"
+    )
+
+    with pytest.raises(Exception) as excinfo:
+        app.run_server(port="65536")
+    assert (
+        excinfo.exconly()
+        == "AssertionError: Expecting an integer from 1 to 65535, found port=65536"
+    )
